Removed unnecessary timer, added server-side logging and added debug texts
I originally thought it would be good to delay the test upon connection since some players might take longer until they finish loading. This plugin is just as reliable if we remove the timer.
Added a ServerLog call to enable server admins to search for filenames that are regularly kicked (in case the filelist becomes obsolete)
Added debug functionality for local feedback of what's checked and what the results are